+++ frontend/src/utils/loadWallMatrix.ts
@@ -1,5 +1,5 @@
-export const loadWallMatrix = async (mapTitle: string): Promise<Float64Array> => {
+export const loadWallMatrix = async (mapTitle: string): Promise<Int16Array> => {
 	const response = await fetch(`/maps/${mapTitle}/wallsMatrix3D_processed.bin`);
 	const buffer = await response.arrayBuffer();
-	return new Float64Array(buffer);
+	return new Int16Array(buffer);
 };
